[Accessibility] Spectrum/focus-ring-polyfill to control display of focus ring depending on interaction mode (#95)
Spectrum-css uses .focus-ring class to indicate that an element
should display a focus style when receiving focus following keyboard
navigation. Javascript must be added to the framework to listen for
keyboard, mouse, focus and blur events on the document and add or
remove the `.focus-ring` class on the `document.activeElement`
depending on the manner of interaction.

Including https://git.corp.adobe.com/Spectrum/focus-ring-polyfill makes
this easy.

Also, add focusRing decorator to ensure that element with `focus-ring`
className doesn’t lose it should component render.